Improve area persistence E2E tests per review feedback
- Extract area helpers to tests/e2e/helpers/areas.js for reuse
- Import shared unique and waitForApp from helpers/todos.js
- Replace all waitForTimeout calls with condition-based waits:
  - selectArea now waits for label to update
  - delete/Shift+0 tests use toContainText assertion instead of delay
- waitForApp now uses loadingScreen check instead of 2s fixed delay
